Separate IAbstractClrNativeHeaps
