+++ src/Microsoft.Diagnostics.Runtime/ClrModule.cs
@@ -20,7 +20,7 @@ namespace Microsoft.Diagnostics.Runtime
     public sealed class ClrModule : IClrModule
     {
         private readonly IAbstractModuleProvider? _helpers;
-        private readonly IAbstractNativeHeapProvider? _nativeHeapHelpers;
+        private readonly IAbstractClrNativeHeaps? _nativeHeapHelpers;
         private int _debugMode = int.MaxValue;
         private MetadataImport? _metadata;
         private PdbInfo? _pdb;
@@ -33,7 +33,7 @@ namespace Microsoft.Diagnostics.Runtime
         internal ClrHeap Heap => _heap ??= AppDomain.Runtime.Heap;
         internal IDataReader DataReader { get; }
 
-        internal ClrModule(ClrAppDomain domain, in ClrModuleInfo data, IAbstractModuleProvider? moduleHelpers, IAbstractNativeHeapProvider? nativeHeapHelpers, IDataReader dataReader)
+        internal ClrModule(ClrAppDomain domain, in ClrModuleInfo data, IAbstractModuleProvider? moduleHelpers, IAbstractClrNativeHeaps? nativeHeapHelpers, IDataReader dataReader)
         {
             _helpers = moduleHelpers;
             _nativeHeapHelpers = nativeHeapHelpers;
